==> code_for_stats/clean code/1) environmental correlation.Rmd <==
---
title: "environmental correlation"
author: "Emmerson Wilson"
date: "2024-03-01"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  ggplot2,
  vegan,
  terra,
  corrplot,
  PerformanceAnalytics,
  usdm,
  janitor,
  openxlsx,
  raster)
```

```{r}
# change this to whichever folder you downloaded the code to
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats")
```


## import data

### from sites

```{r}
env_sites_GM <- read_xlsx("data required/environmental data/env_sites_GM_final.xlsx") %>% # environmental data from Gros Morne sites
  rename(CC = "CC10fill") # change name of canopy cover column to match others

env_sites_TN <- read_xlsx("data required/environmental data/env_sites_TN_final.xlsx") %>% #environmental data from Terra Nova sites
    rename(CC = "CC10fill") # change name of canopy cover column to match others

```

#### prep
##### remove exclosures and unused columns
```{r}
env_sites <- dplyr::bind_rows(env_sites_TN, env_sites_GM) # put sites together
env_sites$ex_ctrl <- env_sites$ex_ctrl %>% 
  replace_na("pair") # turn na's into character so we can filter out exclosures
env_sites <- env_sites %>% 
  filter(ex_ctrl != "EX") %>% #filter our exclosure sites
  dplyr::select(!ex_ctrl)  %>% # remove exclosure control column
  dplyr::select(!Moose) %>% # remove old moose classification
  dplyr::select(!nitro) # remove nitro because not good dataset
```

##### fix aspect
```{r}
env_sites <- env_sites %>% 
  mutate(ASP = abs(180 - ASP))    # make aspect absolute number of degrees from S

env_sites$ASP <-replace_na(env_sites$ASP, 0.1) # anywhere with slope of 0 has NA aspect, make aspect small
```

##### fill moose TN 05
```{r}
env_sites$Moose_new <- replace_na(env_sites$Moose_new, 3)
```

#### create dataset for each park
```{r}
GM_env_sites <- subset(env_sites, env_sites$park == "GM") # make dataset for just GM
TN_env_sites <- subset(env_sites, env_sites$park == "TN") # make dataset for just TN
```


### from entire parks

```{r}
# bring in raster of each environmental variable
SPC_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/NP_FC_2010_2018/NP_SPC_TN.tif")
names(SPC_TN) <- "SPC_TN"
FAC_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/NP_FC_2010_2018/NP_FAC_TN.tif")
names(FAC_TN) <- "FAC_TN"
EVIamp_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GLANCE30_EVI_2019/GLANCE30_EVIamp_TN.tif")
names(EVIamp_TN) <- "EVIamp_TN"
EVImed_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GLANCE30_EVI_2019/GLANCE30_EVImed_TN.tif")
names(EVImed_TN) <- "EVImed_TN"
FHT_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GFCH_FHT_2019/GFCH_FHT_TN.tif")
names(FHT_TN) <- "FHT_TN"
CC10fill_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GFCC_CC_2005_2010/GFCC_CC_TN_10_filled.tif")
names(CC10fill_TN) <- "CC10fill_TN"
SLO_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CDEM_topography_2011/CDEM_SLO_TN.tif")
names(SLO_TN) <- "SLO_TN"
ELE_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CDEM_topography_2011/CDEM_ELE_TN.tif")
names(ELE_TN) <- "ELE_TN"
ASP_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CDEM_topography_2011/CDEM_ASP_TN.tif")
names(ASP_TN) <- "ASP_TN"
LCT_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CEC_LCT_2020/CEC_LCT_2020_TN.tif")
names(LCT_TN) <- "LCT_TN"
LGS_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/MSLSP_EVI_2019/MSLSP_LGS_TN_filled.tif")
names(LGS_TN) <- "LGS_TN"
moose_TN <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/NP_FC_2010_2018/moose_surveys/TN_2020/NP_moose2020_TN_clip.tif")
names(moose_TN) <- "moose_TN"



SPC_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/NP_FC_2010_2018/NP_SPC_GM.tif")
names(SPC_GM) <- "SPC_GM"
FAC_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/NP_FC_2010_2018/NP_FAC_GM.tif")
names(FAC_GM) <- "FAC_GM"
EVIamp_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GLANCE30_EVI_2019/GLANCE30_EVIamp_GM.tif")
names(EVIamp_GM) <- "EVIamp_GM"
EVImed_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GLANCE30_EVI_2019/GLANCE30_EVImed_GM.tif")
names(EVImed_GM) <- "EVImed_GM"
FHT_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GFCH_FHT_2019/GFCH_FHT_GM.tif")
names(FHT_GM) <- "FHT_GM"
CC15fill_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/GFCC_CC_2005_2010/GFCC_CC_GM_15_filled.tif")
names(CC15fill_GM) <- "CC15fill_GM"
SLO_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CDEM_topography_2011/CDEM_SLO_GM.tif")
names(SLO_GM) <- "SLO_GM"
ELE_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CDEM_topography_2011/CDEM_ELE_GM.tif")
names(ELE_GM) <- "ELE_GM"
ASP_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CDEM_topography_2011/CDEM_ASP_GM.tif")
names(ASP_GM) <- "ASP_GM"
LCT_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/CEC_LCT_2020/CEC_LCT_2020_GM.tif")
names(LCT_GM) <- "LCT_GM"
LGS_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/MSLSP_EVI_2019/MSLSP_LGS_GM_filled.tif")
names(LGS_GM) <- "LGS_GM"
moose_GM <- rast("/Volumes/Emmerson's backup/C_forecasting_NL_env_data/All in NAD 83 - UTM 21/NP_FC_2010_2018/moose_surveys/GM_2017/NP_moose2017_GM_clip.tif")
names(moose_GM) <- "moose_GM"


```

#### prep
##### reorder moose data
```{r}
moose_TN[] = ifelse(moose_TN[]==1, 3, ifelse(moose_TN[]==2, 1, ifelse(moose_TN[]==3, 2, 4))) # reclassify moose values so in ascending order of lowest to highest

moose_GM[] = ifelse(moose_GM[]==1, 2, ifelse(moose_GM[]==2, 1, ifelse(moose_GM[]==3, 3, 1))) # reclassify moose values so in ascending order of lowest to highest
```

##### fix aspect
```{r}
values(ASP_GM) <- abs(180 - values(ASP_GM))

### fix na ASP
ASP_GM[is.na(ASP_GM[])] <- 0.01

# TN
values(ASP_TN) <- abs(180 - values(ASP_TN))

### fix na ASP
ASP_TN[is.na(ASP_TN[])] <- 0.01
```

#### stack rasters

```{r}
TN_env_list <- list(SPC_TN, FAC_TN, FHT_TN, CC10fill_TN, LCT_TN, EVIamp_TN, EVImed_TN, SLO_TN, ELE_TN, ASP_TN, LGS_TN, moose_TN)

TN_env_raster <- rast(TN_env_list)

GM_env_list <- list(SPC_GM, FAC_GM, FHT_GM, CC15fill_GM, LCT_GM, EVIamp_GM, EVImed_GM, SLO_GM, ELE_GM, ASP_GM, LGS_GM, moose_GM)

GM_env_raster <- rast(GM_env_list)
```

#### create single dataset
```{r}
TN_env_raster_df <- terra::as.data.frame(TN_env_raster, xy = TRUE)
for ( col in 1:ncol(TN_env_raster_df)){
    colnames(TN_env_raster_df)[col] <-  sub("_TN*", "", colnames(TN_env_raster_df)[col])
} # first need to make raster stack into dataframe

GM_env_raster_df <- as.data.frame(GM_env_raster, xy = TRUE)
for ( col in 1:ncol(GM_env_raster_df)){
    colnames(GM_env_raster_df)[col] <-  sub("_GM*", "", colnames(GM_env_raster_df)[col]) 
} # first need to make raster stack into dataframe

GM_env_raster_df <- GM_env_raster_df %>% 
  rename(CC = "CC15fill")
TN_env_raster_df <- TN_env_raster_df %>% 
  rename(CC = "CC10fill")

NL_env_raster_df <- dplyr::bind_rows(GM_env_raster_df, TN_env_raster_df)
```


## correlation

categorical variables already dummy variables. Moose and SPC made into dummy variables by the string rasterization function in QGIS. and LCT and FAC come in numeric form

#### both parks
```{r, warning=FALSE}
predictors_site_env <- env_sites[, -c(1:7)]
# correlation structure
varCor <- cor(predictors_site_env, use="pairwise.complete.obs")
corrplot(varCor,  tl.col='black') # dendogram
chart.Correlation(predictors_site_env, histogram=F, pch=19)
```

#### at each park

##### GM only

```{r, warning=FALSE}

predictors_site_env_GM <- GM_env_sites[, -c(1:7)]
# correlation structure
varCor_GM <- cor(predictors_site_env_GM, use="pairwise.complete.obs")
corrplot(varCor_GM,  tl.col='black')
chart.Correlation(predictors_site_env_GM, histogram=F, pch=19)
```


##### TN only

```{r, warning=FALSE}

predictors_site_env_TN <- TN_env_sites[,-c(1:7)]
# correlation structure
varCor_TN <- cor(predictors_site_env_TN, use="pairwise.complete.obs")
corrplot(varCor_TN,  tl.col='black')
as.data.frame(varCor_TN)
chart.Correlation(predictors_site_env_TN, histogram=F, pch=19)
```

## rasters

#### both pakrs

```{r, warning=FALSE}
predictors_park_env_NL <- NL_env_raster_df[, c(7,5,6,11,10,9,4,8)]
# correlation structure
# varCor_NL <- cor(predictors_park_env_NL, use="pairwise.complete.obs")
# corrplot(varCor_NL,  tl.col='black')

chart.Correlation(predictors_park_env_NL, histogram=F, pch=19)
```

#### at each park

##### GM only

```{r, warning=FALSE}
predictors_park_env_GM <- GM_env_raster_df[, c(7,5,6,11,10,9,4,8)]
# correlation structure
# varCor_park_GM <- cor(predictors_park_env_GM, use="pairwise.complete.obs")
# corrplot(varCor_park_GM,  tl.col='black')
chart.Correlation(predictors_park_env_GM, histogram=F, pch=19)
```

##### TN only

```{r, warning=FALSE}

predictors_park_env_TN <- TN_env_raster_df[, c(7,5,6,11,10,9,4,8)]
# correlation structure
# varCor_park_TN <- cor(predictors_park_env_TN, use="pairwise.complete.obs")
# corrplot(varCor_park_TN,  tl.col='black')
chart.Correlation(predictors_park_env_TN, histogram=F, pch=19)
```

## export

### for sites
```{r}
write.xlsx(env_sites, file = "data created/env_sites_final.xlsx")
```


### for rasters
export the stacked rasters for each park, because this is what will be used to predict later
```{r}
writeRaster(GM_env_raster, filename="data created/GM_env_raster.tif", overwrite=TRUE)
writeRaster(TN_env_raster, filename="data created/TN_env_raster.tif", overwrite=TRUE)
```
