==> code_for_stats/clean code/2) fill in missing soil c.Rmd <==
---
title: "fill in missing soil c"
author: "Emmerson Wilson"
date: "2024-03-04"
output: html_document
---

all soil sample burned up for 35, one for 41
creating values based on mean soil carbon for other subplots with same park and disturbance type: mature plot in GM

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  janitor)
```

```{r}
# change this to whichever folder you downloaded the code to
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/clean code")
```

## bring in data

### carbon pools for each site
```{r}
c_2023_missing <- read_xlsx("data required/c_data_2023/carbon_data_2023.xlsm", sheet = "Overall C") %>%
  replace(is.na(.), 0) # make each NA a 0
```

### site metainformation
```{r}
site_data <- read_xlsx("data required/c_data_2023/carbon_data_2023.xlsm", sheet = "Subplot info")
site_data$Site_ID <- sprintf("%02.f", site_data$Site_ID) # make names match c data name column
```

### join c and site data
```{r}
c_2023_missing_jn1 <- c_2023_missing %>%
  left_join(site_data)%>%
  clean_names() %>%
  mutate(gap = disturbance) %>% # make a column of wether site gap or mature
  mutate(across('gap', str_replace, 'Insect', 'Gap')) %>% # add if gap or not gap
  mutate(across('gap', str_replace, 'Fire', 'Gap')) %>%
  mutate(across('gap', str_replace, 'Wind', 'Gap')) %>%
  mutate(across('gap', str_replace, 'Logging', 'Gap')) %>%
  mutate(year = 2023) %>% # add year
  mutate(site_id = paste0(site_id, "_", "2023")) %>% 
  filter(ex_ctrl != "EX") # remove exclosures
  

c_2023_missin_jn2 <- c_2023_missing_jn1 %>%
  dplyr::select(c(22,23,25,2,27,38,39,5,7,10,11,13,16,18,21)) %>% # select columns want to use
  replace(is.na(.), 0) %>%
  mutate(unique_id = paste0(park_id, "_", site_id, "_", subplot_id)) # create a unique id for eahc subplot
```


## fill in values for 35
```{r}
c_23_matGM <- subset(c_2023_missin_jn2, park_id == "GM" & disturbance == "Mature")

c_23_matGM_avg <- mean(subset(c_23_matGM, sl_carbon_9m2!= 0)$sl_carbon_9m2)
c_23_matGM_sd <- sd(subset(c_23_matGM, sl_carbon_9m2!= 0)$sl_carbon_9m2)/sqrt(length(subset(c_23_matGM, sl_carbon_9m2!= 0)$sl_carbon_9m2))
c_23_matGM_med <- median(subset(c_23_matGM, sl_carbon_9m2!= 0)$sl_carbon_9m2)

c_2023_fill <- c_2023_missin_jn2

c_2023_fill$sl_carbon_9m2[c_2023_fill$sl_carbon_9m2 == 0] <- c_23_matGM_med

c_2023_fill <- c_2023_fill %>% 
  dplyr::select(!(unique_id)) # will add back in next script
```

### export
```{r}
write.xlsx(c_2023_fill, file = "data created/carbon_data_2023_scfill.xlsx")
```
