==> code_for_stats/clean code/3) c per m2.Rmd <==
---
title: "C per m2"
author: "Emmerson Wilson"
date: "2024-03-04"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  janitor)
```

```{r}
# change this to whichever folder you downloaded the code to
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/clean code")
```

# bring in data

## 2022

```{r}
# only bring in columns needed (ie subplot id and c columns)
c_2022_dw <- read_xlsx("data required/c_data_2022/Deadwood_Field_Data NEW.xlsm", sheet = "DW carbon") %>%
  dplyr::select(c(Site_Subplot_ID, dw_carbon_9m2))
c_2022_te <- read_xlsx("data required/c_data_2022/Tree_Biomass_NEW.xlsm", sheet = "Tree carbon") %>%
  dplyr::select(c(Site_Subplot_ID, te_carbon_9m2))
c_2022_gc <- read_xlsx("data required/c_data_2022/ground_cover_carbon.xlsx", sheet = "Sheet 1") %>%
  dplyr::select(c(Site_Subplot_ID, gc_carbon_9m2))
c_2022_hb <- read_xlsx("data required/c_data_2022/herbaceous_carbon.xlsx", sheet = "Sheet 1") %>%
  dplyr::select(c(Site_Subplot_ID, hb_carbon_9m2))
c_2022_ll <- read_xlsx("data required/c_data_2022/litter_carbon.xlsx", sheet = "Sheet 1") %>%
  dplyr::select(c(Park_ID, Site_ID, Ex_Ctrl, Subplot_ID, Site_Subplot_ID, ll_carbon_9m2))
c_2022_sp <- read_xlsx("data required/c_data_2022/sapling_carbon.xlsx", sheet = "Sheet 1") %>% dplyr::select(c(Site_Subplot_ID, sp_carbon_9m2))
c_2022_sl <- read_xlsx("data required/c_data_2022/soil_carbon.xlsx", sheet = "Sheet 1") %>%
  dplyr::select(c(Site_Subplot_ID, sl_carbon_9m2))
c_2022_tr <- read_xlsx("data required/c_data_2022/Tree_Biomass_NEW.xlsm", sheet = "Tree root carbon") %>%
  dplyr::select(c(Site_Subplot_ID, tr_carbon_9m2))
c_2022_u3 <- read_xlsx("data required/c_data_2022/under_30_carbon.xlsx", sheet = "Sheet 1") %>%
  mutate(u3_carbon_9m2 = u3_carbon_9m2/1000000) %>% # convert units
  dplyr::select(c(Site_Subplot_ID, u3_carbon_9m2))
c_2022_ws <- read_xlsx("data required/c_data_2022/woody_shrubs_carbon.xlsx", sheet = "Sheet 1") %>%
  dplyr::select(c(Site_Subplot_ID, ws_carbon_9m2))
site_data_2022 <- read_xlsx("data required/c_data_2022/subplot_disturbacne.xlsx") # metadata
```

### prep
#### get carbon in from woody shrubs, saplings, under 30 saplings to get root carbon
```{r}
c_2022_sr_list <- list(c_2022_u3, c_2022_ws, c_2022_sp)
c_2022_sr <- c_2022_sr_list %>% reduce(full_join, by='Site_Subplot_ID') %>% 
  replace(is.na(.), 0)
c_2022_sr$sr_carbon_9m2 <- rowSums(c_2022_sr[, c(2:4)])
c_2022_sr <- c_2022_sr %>% dplyr::select(c(1,5)) %>%
  mutate(sr_carbon_9m2 = sr_carbon_9m2*0.25)
```

#### combine dataframes
and select columns using
```{r}
c_2022_list <- list(c_2022_dw, c_2022_te, c_2022_gc, c_2022_hb, c_2022_ll, c_2022_sp, c_2022_sr, c_2022_sl, c_2022_tr, c_2022_u3, c_2022_ws)

c_2022 <- c_2022_list %>% reduce(full_join, by='Site_Subplot_ID') %>%
  dplyr::select(c(1,6:9,2:5,10:16)) %>%
  replace(is.na(.), 0)
```

#### join with metadata
```{r}
c_2022_mess <- c_2022 %>%
  left_join(site_data_2022) %>% # add subplot info
  clean_names() %>% 
  mutate(gap = disturbance) %>% 
  mutate(across('gap', str_replace, 'Insect', 'Gap')) %>% # add if gap or not gap
  mutate(across('gap', str_replace, 'Fire', 'Gap')) %>%
  mutate(ss_carbon_9m2 = rowSums(c_2022[, c("sp_carbon_9m2", "ws_carbon_9m2")])) %>% # recategorize pools to match 2023, here create shrub and sapling pool
  mutate(rt_carbon_9m2 = rowSums(c_2022[, c("tr_carbon_9m2","sr_carbon_9m2")])) %>% # combine tree and other roots
  mutate(gv_carbon_9m2 = rowSums(c_2022[, c("gc_carbon_9m2", "hb_carbon_9m2", "u3_carbon_9m2")])) %>% # quadrat c
  mutate(year = 2022) %>% # add year
  mutate(site_number = as.numeric(site_id))
c_2022_mess$site_number <- sprintf("%02.f", c_2022_mess$site_number) # fix site id
c_2022_mess <- c_2022_mess %>%
  mutate(site_id = paste0(site_number, "_", "2022")) %>% # distinguish overlapping site numbers by year
  mutate(subplot_id = substr(subplot_id,2,2)) %>% # remove letter from subplot
  filter(ex_ctrl != "EX")
```

#### slect columns
that will be in common with final 2023 dataframe
```{r}
c_2022_compat <- c_2022_mess %>%
  dplyr::select(c(2:5,17,18,22,6,7,10,13,19:21))
```

#### make new total c
```{r}
c_2022_compat$total_c_per_subplot <- rowSums(c_2022_compat[, c(8:14)])
```

## 2023

```{r}
c_2023 <- read_xlsx("data created/carbon_data_2023_scfill.xlsx") %>% # carbon data (soil filled in)
  replace(is.na(.), 0) 
```

### prep
#### join metadata
was already joined with metadata while adding missing soil c

#### select columns
that are common with final 2022 dataframe
was done while adding missing soil c

## combine

```{r}
c_22_23 <- rbind(c_2022_compat, c_2023) %>%
  mutate(unique_id = paste0(park_id, "_", site_id, "_", subplot_id)) # add unique id

write.xlsx(c_22_23, file = "data created/c_22_23.xlsx")
```

# total carbon per 1m^2 
need to average carbon at each subplot within a site
```{r}
c_22_23_m2 <- c_22_23 %>%
  mutate(unique_id = paste0(park_id, "_",site_id)) %>% 
  group_by(unique_id, park_id, gap) %>% # group by id, and still want park_id and gao columns in output
  summarize(c_m2_site = sum(total_c_per_subplot)/36) # divide by 36 because each of the 4 subplots is 9m2

c_m2_site <- write.xlsx(c_22_23_m2, file = "data created/c_m2_site.xlsx")
```
