==> code_for_stats/clean code/12) c error propagation.Rmd <==
---
title: "12) c error propagation"
author: "Emmerson Wilson"
date: "2024-05-08"
output: html_document
---
```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  terra)

```

```{r}
# change this to whichever folder you downloaded the code to
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/clean code")
```

# import training datasets
```{r}
# made in 5) c lm total.Rmd
GM_c_env <- read_xlsx("data created/c_env_sites_scale_GM.xlsx")
TN_c_env <- read_xlsx("data created/c_env_sites_scale_TN.xlsx")

mean(GM_c_env$c_m2_site)/1000
mean(TN_c_env$c_m2_site)/1000


```

# bring in files
```{r}
GM_env_raster_masked_scale_bm <- rast("data created/GM_env_raster_masked_scale_bm_nwl.tif")
TN_env_raster_masked_scale_bm <- rast("data created/TN_env_raster_masked_scale_bm_nwl.tif")
```


# GM
## make predictions using top model and stacked environmental rasters, transformed onto scale of dependant variable

## model
```{r}
# list of variables in best model from 5) c lm total.Rmd
GM_var <- list("FHT", "EVIamp")

# run best model
GM_model_tot <- lm_glm(GM_var, df = GM_c_env, c = "c_m2_site")
summary(GM_model_tot)
```


need to run 8) c predictions first
```{r}
GM_predict_tot1 <- terra::predict(GM_env_raster_masked_scale_bm, GM_model_tot, se.fit = T)
GM_predict_tot <- exp(GM_predict_tot1)
```

## get standard errors of the predicted mean from predict() function, transform to scale of dependent variable 
```{r}
GM_predict_tot_se1 <- GM_predict_tot1$se.fit
GM_predict_tot_se <- exp(GM_predict_tot_se1)

# export
writeRaster(GM_predict_tot_se, filename="data created/GM_predict_tot_se.tif", overwrite=TRUE)
```

## calculate CI from se of fit
```{r}
GM_CI_upper_95 <- exp(GM_predict_tot1$fit + 1.96*GM_predict_tot_se1)
GM_CI_lower_95 <- exp(GM_predict_tot1$fit - 1.96*GM_predict_tot_se1)
```


## get standard errors from residual variance (of prediction interval), transfomr to scale of dependant variable
```{r}
GM_se.PI1 <- sqrt(GM_predict_tot_se1$se.fit ^ 2 + GM_predict_tot1$residual.scale ^ 2)
GM_se.PI <- exp(GM_se.PI1)

# export
writeRaster(GM_se.PI, filename="data created/GM_predict_tot_pse.tif", overwrite=TRUE)
```

## calculate PI from se fo residual variance
```{r}
GM_PI_upper_95 <- exp(GM_predict_tot1$fit + 1.96*GM_se.PI1)
GM_PI_lower_95 <- exp(GM_predict_tot1$fit - 1.96*GM_se.PI1)
```

## plots

### se's and residuals violin plots
```{r}
GM_se_s <- list(se.CI = GM_predict_tot_se, se.PI = GM_se.PI, resid = GM_predict_tot1$residual.scale)

GM_se_s_df <- lapply(names(GM_se_s), function(i) {
  data.frame(
    rastername = i,
    value = as.vector(GM_se_s[[i]])
  )
})

GM_se_s_df <- do.call(rbind, GM_se_s_df)

GM_se_s_df %>% 
  group_by(rastername) %>% 
  summarize(max = max(value, na.rm = T), min = min(value, na.rm = T))

ggplot(GM_se_s_df, aes(rastername, value, fill = rastername)) +
  geom_violin() 
```

### se's
```{r}
plot(GM_predict_tot_se)
plot(GM_se.PI)
```

### CI
```{r}
plot(GM_CI_upper_95)
plot(GM_CI_lower_95)
```

### PI
```{r}
plot(GM_PI_upper_95)
plot(GM_PI_lower_95)
```

# TN
## make predictions using top model and stacked environmental rasters, transformed onto scale of dependant variable

## mdoel
```{r}
# list of variables in best model from 5) c lm total.Rmd
TN_var <- list("FHT", "EVIamp")

# run best model
TN_model_tot <- lm_glm(TN_var, df = subset(TN_c_env), c = "c_m2_site")
summary(TN_model_tot)
```



```{r}
TN_predict_tot1 <- terra::predict(TN_env_raster_masked_scale_bm, TN_model_tot, se.fit = T)
TN_predict_tot <- exp(TN_predict_tot1)
```

## get standard errors of the predicted mean from predict() function, transform to scale of dependent variable 
```{r}
TN_predict_tot_se1 <- TN_predict_tot1$se.fit
TN_predict_tot_se <- exp(TN_predict_tot_se1)

# export
writeRaster(TN_predict_tot_se, filename="data created/TN_predict_tot_se.tif", overwrite=TRUE)

```

## calculate CI from se of fit
```{r}
TN_CI_upper_95 <- exp(TN_predict_tot1$fit + 1.96*TN_predict_tot_se1)
TN_CI_lower_95 <- exp(TN_predict_tot1$fit - 1.96*TN_predict_tot_se1)
```


## get standard errors from residual variance (of prediction interval), transfomr to scale of dependant variable
```{r}
TN_se.PI1 <- sqrt(TN_predict_tot_se1$se.fit ^ 2 + TN_predict_tot1$residual.scale ^ 2)
TN_se.PI <- exp(TN_se.PI1)

#export
writeRaster(TN_se.PI, filename="data created/TN_predict_tot_pse.tif", overwrite=TRUE)
```

## calculate PI from se fo residual variance
```{r}
TN_PI_upper_95 <- exp(TN_predict_tot1$fit + 1.96*TN_se.PI1)
TN_PI_lower_95 <- exp(TN_predict_tot1$fit - 1.96*TN_se.PI1)
```

## plots

### se's and residuals violin plots
```{r}
TN_se_s <- list(se.CI = TN_predict_tot_se, se.PI = TN_se.PI, resid = TN_predict_tot1$residual.scale)

TN_se_s_df <- lapply(names(TN_se_s), function(i) {
  data.frame(
    rastername = i,
    value = as.vector(TN_se_s[[i]])
  )
})

TN_se_s_df <- do.call(rbind, TN_se_s_df)

TN_se_s_df %>% 
  group_by(rastername) %>% 
  summarize(max = max(value, na.rm = T), min = min(value, na.rm = T))

ggplot(TN_se_s_df, aes(rastername, value, fill = rastername)) +
  geom_violin() 
```

### se's
```{r}
plot(TN_predict_tot_se)
plot(TN_se.PI)
plot(TN_resid)
```

### CI
```{r}
plot(TN_CI_upper_95)
plot(TN_CI_lower_95)
```

### PI
```{r}
plot(TN_PI_upper_95)
plot(TN_PI_lower_95)
```
