==> code_for_stats/clean code/13) gap error propagation.Rmd <==
---
title: "13) gap error propagation"
author: "Emmerson Wilson"
date: "2024-05-08"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  terra,
  brglm2)

```

```{r}
# change this to whichever folder you downloaded the code to
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/clean code")
```

# import training datasets
```{r}
# made in 5) c lm total.Rmd
GM_gap_env <- read_xlsx("data created/c_env_sites_scale_GM_gap.xlsx")
TN_gap_env <- read_xlsx("data created/c_env_sites_scale_TN_gap.xlsx")
```

## for space based on variabels in best models
```{r}
GM_env_raster_masked_scaled_bm <- rast("data created/GM_env_raster_masked_scale_bm_nwl.tif") # stack of enviornmental variables' rasters
TN_env_raster_masked_scaled_bm <- rast("data created/TN_env_raster_masked_scale_bm_nwl.tif")
```

# import glm function
```{r}
source("4) glm function.R")
```

# GM
## make predictions using top model and stacked environmental rasters, transformed onto scale of dependant variable
```{r}
GM_var_gap <- list("FHT", "EVIamp")

# run best model
GM_model_tot_gap <- firths_glm(GM_var_gap, df = GM_gap_env, c = "gap")

GM_predict_tot_gap1 <- terra::predict(GM_env_raster_masked_scaled_bm, GM_model_tot_gap, se.fit = T)
GM_predict_tot_gap <- 1 / (1 + exp(-GM_predict_tot_gap1))
```

## get standard errors of the predicted mean from predict() function, transform to scale of dependent variable 
```{r}
GM_predict_tot_gap_se1 <- GM_predict_tot_gap1$se.fit
GM_predict_tot_gap_se <- 1 / (1 + exp(-GM_predict_tot_gap_se1))

#export
writeRaster(GM_predict_tot_gap_se, filename="data created/GM_predict_tot_gap_se.tif", overwrite=TRUE)
```

## calculate CI from se of fit
```{r}
GM_CI_upper_95_gap <- 1 / (1 + exp(-(GM_predict_tot_gap1$fit + 1.96*GM_predict_tot_gap_se1)))
GM_CI_lower_95_gap <- 1 / (1 + exp(-(GM_predict_tot_gap1$fit - 1.96*GM_predict_tot_gap_se1)))


```


## get standard errors from residual variance (of prediction interval), transfomr to scale of dependant variable
```{r}
GM_se.PI_gap1 <- sqrt(GM_predict_tot_gap_se1$se.fit ^ 2 + GM_predict_tot_gap1$residual.scale ^ 2)
GM_se.PI_gap <- 1 / (1 + exp(-GM_se.PI_gap1))

writeRaster(GM_se.PI_gap, filename="data created/GM_predict_tot_gap_pse.tif", overwrite=TRUE)

```

## calculate PI from se fo residual variance
```{r}
GM_PI_upper_95_gap <- 1 / (1 + exp(-(GM_predict_tot_gap1$fit + 1.96*GM_se.PI_gap1)))
GM_PI_lower_95_gap <- 1 / (1 + exp(-(GM_predict_tot_gap1$fit - 1.96*GM_se.PI_gap1)))
```

## plots

### se's and residuals violin plots
```{r}

GM_se_s_gap <- list(se.CI = GM_predict_tot_gap_se, se.PI = GM_se.PI_gap, resid = GM_predict_tot_gap1$residual.scale)

GM_se_s_df_gap <- lapply(names(GM_se_s_gap), function(i) {
  data.frame(
    rastername = i,
    value = as.vector(GM_se_s_gap[[i]])
  )
})

GM_se_s_df_gap <- do.call(rbind, GM_se_s_df_gap)

GM_se_s_df_gap %>% 
  group_by(rastername) %>% 
  summarize(max = max(value, na.rm = T), min = min(value, na.rm = T))

ggplot(GM_se_s_df_gap, aes(rastername, value, fill = rastername)) +
  geom_violin() 



```

### se's
```{r}
plot(GM_predict_tot_gap_se)
plot(GM_se.PI_gap)
plot(GM_resid_gap)
```

### CI
```{r}
plot(GM_CI_upper_95_gap)
plot(GM_CI_lower_95_gap)
```

### PI
```{r}
plot(GM_PI_upper_95_gap)
plot(GM_PI_lower_95_gap)
```

# TN
## make predictions using top model and stacked environmental rasters, transformed onto scale of dependant variable
```{r}
TN_var_gap <- list("FHT", "EVIamp")

# run best model
TN_model_tot_gap <- firths_glm(TN_var_gap, df = TN_gap_env, c = "gap")

TN_predict_tot_gap1 <- terra::predict(TN_env_raster_masked_scaled_bm, TN_model_tot_gap, se.fit = T)
TN_predict_tot_gap <- 1 / (1 + exp(-TN_predict_tot_gap1))
```

## get standard errors of the predicted mean from predict() function, transform to scale of dependent variable 
```{r}
TN_predict_tot_gap_se1 <- TN_predict_tot_gap1$se.fit
TN_predict_tot_gap_se <- 1 / (1 + exp(-TN_predict_tot_gap_se1))

writeRaster(TN_predict_tot_gap_se, filename="data created/TN_predict_tot_gap_se.tif", overwrite=TRUE)

```

## calculate CI from se of fit
```{r}
TN_CI_upper_95_gap <- 1 / (1 + exp(-(TN_predict_tot_gap1$fit + 1.96*TN_predict_tot_gap_se1)))
TN_CI_lower_95_gap <- 1 / (1 + exp(-(TN_predict_tot_gap1$fit - 1.96*TN_predict_tot_gap_se1)))
```


## get standard errors from residual variance (of prediction interval), transfomr to scale of dependant variable
```{r}
TN_se.PI_gap1 <- sqrt(TN_predict_tot_gap_se1$se.fit ^ 2 + TN_predict_tot_gap1$residual.scale ^ 2)
TN_se.PI_gap <- 1 / (1 + exp(-TN_se.PI_gap1))

writeRaster(TN_se.PI_gap, filename="data created/TN_predict_tot_gap_pse.tif", overwrite=TRUE)
```

## calculate PI from se fo residual variance
```{r}
TN_PI_upper_95_gap <- 1 / (1 + exp(-(TN_predict_tot_gap1$fit + 1.96*TN_se.PI_gap1)))
TN_PI_lower_95_gap <- 1 / (1 + exp(-(TN_predict_tot_gap1$fit - 1.96*TN_se.PI_gap1)))
```

## plots

### se's and residuals violin plots
```{r}

TN_se_s_gap <- list(se.CI = TN_predict_tot_gap_se, se.PI = TN_se.PI_gap, resid =TN_predict_tot_gap1$residual.scale)

TN_se_s_df_gap <- lapply(names(TN_se_s_gap), function(i) {
  data.frame(
    rastername = i,
    value = as.vector(TN_se_s_gap[[i]])
  )
})

TN_se_s_df_gap <- do.call(rbind, TN_se_s_df_gap)

TN_se_s_df_gap %>% 
  group_by(rastername) %>% 
  summarize(max = max(value, na.rm = T), min = min(value, na.rm = T))

ggplot(TN_se_s_df_gap, aes(rastername, value, fill = rastername)) +
  geom_violin() 
```

### se's
```{r}
plot(TN_predict_tot_gap_se)
plot(TN_se.PI_gap)
plot(TN_resid_gap)
```

### CI
```{r}
plot(TN_CI_upper_95_gap)
plot(TN_CI_lower_95_gap)
```

### PI
```{r}
plot(TN_PI_upper_95_gap)
plot(TN_PI_lower_95_gap)
```
