==> code_for_stats/clean code/20) Exporting maps for figures.Rmd <==
---
title: "Exporting maps for figures"
author: "Emmerson Wilson"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(leaflet)
library(leafem)
library(mapview)
library(webshot2)

setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/clean code")
```

# data
```{r}
# roads and park boundaries
roads <- vect("data required/environmental data/parks and roads/highways_nl.shp") %>% 
  project("EPSG:4326 - WGS 84")

parksGM <- vect("data required/environmental data/parks and roads/GM_outline.shp") %>% 
  project("EPSG:4326 - WGS 84")
parksTN <- vect("data required/environmental data/parks and roads/TN_outline.shp") %>% 
  project("EPSG:4326 - WGS 84")

# Gros Morne data and palettes
GM_c <- rast("data created/GM_predict_tot_bm.tif")
GM_c_mean <- rast("data created/GM_predict_tot_se.tif")
GM_c_pred <- rast("data created/GM_predict_tot_pse.tif")

GM_g <- rast("data created/GM_predict_gap_bm.tif")
GM_g_mean <- rast("data created/GM_predict_tot_gap_se.tif")
GM_g_pred <- rast("data created/GM_predict_tot_gap_pse.tif")

# Terra Nova
TN_c <- rast("data created/TN_predict_tot_bm.tif")
TN_c_mean <- rast("data created/TN_predict_tot_se.tif")
TN_c_pred <- rast("data created/TN_predict_tot_pse.tif")

TN_g <- rast("data created/TN_predict_gap_bm.tif")
TN_g_mean <- rast("data created/TN_predict_tot_gap_se.tif")
TN_g_pred <- rast("data created/TN_predict_tot_gap_pse.tif")


```

# colours
```{r include=FALSE}
GMc.pal <- colorNumeric(c("#faf6ef", "#ed6f66", "#8a1c18", "#6c100d"), domain = rev(values(GM_c)), na.color = "transparent")
GMcmean.pal <- colorNumeric(c("white", "#6a51a3", "#3f007d"), domain = rev(values(GM_c_mean)), na.color = "transparent")
GMcpred.pal <- colorNumeric(c("white", "#2171b5", "#092955"), domain = rev(values(GM_c_pred)), na.color = "transparent")

GMg.pal <- colorNumeric(c("#faf6ef", "#f7eedf", "#de5714"), domain = rev(0:1), na.color = "transparent")
GMgmean.pal <- colorNumeric(c("white", "#6a51a3", "#3f007d"), domain = rev(values(GM_g_mean)), na.color = "transparent")
GMgpred.pal <- colorNumeric(c("white", "#2171b5", "#092955"), domain = rev(values(GM_g_pred)), na.color = "transparent")

TNc.pal <- colorNumeric(c("#faf6ef", "#ed6f66", "#8a1c18", "#6c100d"), domain = rev(values(TN_c)), na.color = "transparent")
TNcmean.pal <- colorNumeric(c("white", "#6a51a3", "#3f007d"), domain = rev(values(TN_c_mean)), na.color = "transparent")
TNcpred.pal <- colorNumeric(c("white", "#2171b5", "#092955"), domain = rev(values(TN_c_pred)), na.color = "transparent")

TNg.pal <- colorNumeric(c("#faf6ef", "#f7eedf", "#de5714"), domain = rev(0:1), na.color = "transparent")
TNgmean.pal <- colorNumeric(c("white", "#6a51a3", "#3f007d"), domain = rev(values(TN_g_mean)), na.color = "transparent")
TNgpred.pal <- colorNumeric(c("white", "#2171b5", "#092955"), domain = rev(values(TN_g_pred)), na.color = "transparent")
```

#GM
## carbon

### estimates
```{r}
GM_map_c <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksGM, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = GM_c , 
                     color = GMc.pal,
                     opacity = 1,
                     group = "Estimated C stocks",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = GMc.pal, values = values(GM_c),
                title = "Total carbon (g/m<sup>2</sup>)",
                group = "Estimated C stocks", opacity = 1, position = "topleft") %>% 
  fitBounds(-57.72 ,49.37, -57.82, 49.87) %>%
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

GM_map_c

mapviewOptions(fgb = FALSE)
Sys.setenv(CHROMOTE_CHROME = "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome")

mapview::mapshot2(
  GM_map_c,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/GM_c_map.png",
  remove_controls = "zoomControl")
```

### mean error
```{r}
GM_map_cse <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksGM, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = GM_c_mean , 
                     color = GMcmean.pal,
                     opacity = 1,
                     group = "Estimated C stocks",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = GMcmean.pal, values = values(GM_c_mean),
                title = "Mean residual error for </br> total carbon estimates",
                group = "Estimated C stocks", opacity = 1, position = "topleft") %>% 
  fitBounds(-57.72 ,49.37, -57.82, 49.87) %>%
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

GM_map_cse

mapview::mapshot2(
  GM_map_cse,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/GM_c_map_se.png",
  remove_controls = "zoomControl")
```

### prediction error
```{r}
GM_map_cpe <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksGM, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = GM_c_pred , 
                     color = GMcpred.pal,
                     opacity = 1,
                     group = "Estimated C stocks",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = GMcpred.pal, values = values(GM_c_pred),
                title = "Prediction error for </br> total carbon estimates",
                group = "Estimated C stocks", opacity = 1, position = "topleft") %>% 
  fitBounds(-57.72 ,49.37, -57.82, 49.87) %>%
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

GM_map_cpe

mapview::mapshot2(
  GM_map_cpe,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/GM_c_map_pe.png",
  remove_controls = "zoomControl")
```
## gap

### estimates
```{r}
GM_map_g <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksGM, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = GM_g , 
                     color = GMg.pal,
                     opacity = 1,
                     group = "Estimated mature forest/gap",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = GMg.pal, values = values(GM_g),
                title = "Probability area is mature forest",
                group = "Estimated mature forest/gap", opacity = 1, position = "topleft") %>% 
  fitBounds(-57.72 ,49.37, -57.82, 49.87) %>%
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

GM_map_g

mapview::mapshot2(
  GM_map_g,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/GM_g_map.png",
  remove_controls = "zoomControl")
```

### mean error
```{r}
GM_map_gse <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksGM, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = GM_g_mean , 
                     color = GMgmean.pal,
                     opacity = 1,
                     group = "Estimated mature forest/gap",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = GMgmean.pal, values = values(GM_g_mean),
                title = "Mean residual error for </br> probability area is mature forest",
                group = "Estimated mature forest/gap", opacity = 1, position = "topleft") %>% 
  fitBounds(-57.72 ,49.37, -57.82, 49.87) %>%
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

GM_map_gse

mapview::mapshot2(
  GM_map_gse,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/GM_g_map_se.png",
  remove_controls = "zoomControl")
```

### prediction error
```{r}
GM_map_gpe <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksGM, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = GM_g_pred , 
                     color = GMgpred.pal,
                     opacity = 1,
                     group = "Estimated mature forest/gap",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = GMgpred.pal, values = values(GM_g_pred),
                title = "Prediction error for </br> probability area is mature forest",
                group = "Estimated mature forest/gap", opacity = 1, position = "topleft") %>% 
  fitBounds(-57.72 ,49.37, -57.82, 49.87) %>%
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

GM_map_gpe

mapview::mapshot2(
  GM_map_gpe,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/GM_g_map_pe.png",
  remove_controls = "zoomControl")
```

#TN
## carbon

### estimates
```{r}
TN_map_c <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksTN, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = TN_c , 
                     color = TNc.pal,
                     opacity = 1,
                     group = "Estimated C stocks",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = TNc.pal, values = values(TN_c),
                title = "Total carbon (g/m<sup>2</sup>)",
                group = "Estimated C stocks", opacity = 1, position = "topleft") %>% 
      fitBounds(-54.1, 48.38, -53.8, 48.66)%>% 
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

TN_map_c

mapviewOptions(fgb = FALSE)
Sys.setenv(CHROMOTE_CHROME = "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome")

mapview::mapshot2(
  TN_map_c,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/TN_c_map.png",
  remove_controls = "zoomControl")
```

### mean error
```{r}
TN_map_cse <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksTN, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = TN_c_mean , 
                     color = TNcmean.pal,
                     opacity = 1,
                     group = "Estimated C stocks",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = TNcmean.pal, values = values(TN_c_mean),
                title = "Mean residual error for </br> total carbon estimates",
                group = "Estimated C stocks", opacity = 1, position = "topleft") %>% 
      fitBounds(-54.1, 48.38, -53.8, 48.66)%>% 
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

TN_map_cse

mapview::mapshot2(
  TN_map_cse,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/TN_c_map_se.png",
  remove_controls = "zoomControl")
```

### prediction error
```{r}
TN_map_cpe <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksTN, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = TN_c_pred , 
                     color = TNcpred.pal,
                     opacity = 1,
                     group = "Estimated C stocks",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = TNcpred.pal, values = values(TN_c_pred),
                title = "Prediction error for </br> total carbon estimates",
                group = "Estimated C stocks", opacity = 1, position = "topleft") %>% 
      fitBounds(-54.1, 48.38, -53.8, 48.66)%>% 
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

TN_map_cpe

mapview::mapshot2(
  TN_map_cpe,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/TN_c_map_pe.png",
  remove_controls = "zoomControl")
```
## gap

### estimates
```{r}
TN_map_g <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksTN, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = TN_g , 
                     color = TNg.pal,
                     opacity = 1,
                     group = "Estimated mature forest/gap",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = TNg.pal, values = values(TN_g),
                title = "Probability area is mature forest",
                group = "Estimated mature forest/gap", opacity = 1, position = "topleft") %>% 
      fitBounds(-54.1, 48.38, -53.8, 48.66)%>% 
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

TN_map_g

mapview::mapshot2(
  TN_map_g,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/TN_g_map.png",
  remove_controls = "zoomControl")
```

### mean error
```{r}
TN_map_gse <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksTN, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = TN_g_mean , 
                     color = TNgmean.pal,
                     opacity = 1,
                     group = "Estimated mature forest/gap",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = TNgmean.pal, values = values(TN_g_mean),
                title = "Mean residual error for </br> probability area is mature forest",
                group = "Estimated mature forest/gap", opacity = 1, position = "topleft") %>% 
      fitBounds(-54.1, 48.38, -53.8, 48.66)%>% 
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

TN_map_gse

mapview::mapshot2(
  TN_map_gse,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/TN_g_map_se.png",
  remove_controls = "zoomControl")
```

### prediction error
```{r}
TN_map_gpe <- leaflet(height = 850, width = 800) %>%
      addProviderTiles('Esri.WorldImagery') %>%
      addMapPane("parks", zIndex = 410) %>% 
      addMapPane("roads", zIndex = 415) %>%
      addMapPane("coutput", zIndex = 435) %>% 
      addPolylines(data = roads, color = "#efbc15", opacity = 1, weight = 1.25,
                   options = pathOptions(pane = "roads")) %>%
      addPolygons(data = parksTN, color = "#1f78b4", weight = 1.25, fillColor = "#1f78b4", fillOpacity = 0.5,
                  options = pathOptions(pane = "parks")) %>% 
      # carbon estimates
      addRasterImage(x = TN_g_pred , 
                     color = TNgpred.pal,
                     opacity = 1,
                     group = "Estimated mature forest/gap",
                     options = pathOptions(pane = "coutput")) %>%
      addLegend(pal = TNgpred.pal, values = values(TN_g_pred),
                title = "Prediction error for </br> probability area is mature forest",
                group = "Estimated mature forest/gap", opacity = 1, position = "topleft") %>% 
      fitBounds(-54.1, 48.38, -53.8, 48.66)%>% 
  addScaleBar(position = "bottomright", options = scaleBarOptions(imperial = FALSE, maxWidth = 200))

TN_map_gpe

mapview::mapshot2(
  TN_map_gpe,
  file = "~/Documents/Master's/C_forecasting_NL/code_for_stats/clean code/data created/map figrues/TN_g_map_pe.png",
  remove_controls = "zoomControl")
```
